Test enhancement
