+++ composer.json
@@ -22,7 +22,7 @@
         "ext-json": "*"
     },
     "require-dev": {
-        "phpunit/phpunit" : ">=5.4.3",
+        "phpunit/phpunit" : ">=6.5",
         "squizlabs/php_codesniffer": "^2.3"
     },
     "autoload": {
