+++ src/Details.php
@@ -7,11 +7,11 @@ namespace ipinfo\ipinfo;
  */
 class Details
 {
-  public function __construct($raw_details)
-  {
-    foreach ($raw_details as $property => $value) {
-      $this->$property = $value;
+    public function __construct($raw_details)
+    {
+        foreach ($raw_details as $property => $value) {
+            $this->$property = $value;
+        }
+        $this->all = $raw_details;
     }
-    $this->all = $raw_details;
-  }
 }
